feat: Added first attempt at Vue based filter builder
First attempt at a Vue component as well, so major refactoring is expected.  The PHP side criteria builder is not fully implemented and the filter data is not consistently being emitted up the tree (some rely on a the prop containing proxies). There are also some bugs related to deleting conditions (the wrong item seems to be removed from the DOM, but not the internal array).